+++ python/tank/descriptor/__init__.py
@@ -21,5 +21,5 @@ from .errors import (
     TankInvalidInterpreterLocationError, TankMissingManifestError
 )
 
-from .descriptor import create_descriptor
+from .descriptor_factory import create_descriptor
 from .io_descriptor import descriptor_dict_to_uri, descriptor_uri_to_dict, is_descriptor_version_missing
